// src/services/iegm/types.ts
export interface IEGMFilters {
  ano?: number;
  tribunal?: string;
  municipio?: string;
  indicador?: string;
  faixa?: string;
}